java: Use JUnit BOM to manage JUnit5 dependencies (#1748)
Unlike JUnit4, JUnit5 is a multi-module test framework. The version of these
modules should be kept in sync. This is best managed by using the JUnit Bill
of Materials.
When used in the dependency management section this allows individual JUnit
modules to be used without a version. They'll implicitly have the version
specified in the BOM.

By having a single version this should also ensure that tools like Renovate
will update all versions consistently.